Rename: 패키지 정리를 통한 프로젝트의 패키지 구조 개선
