+++ queosk/src/main/java/com/bttf/queosk/config/baseEntity/JpaAuditingConfiguration.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.config;
+package com.bttf.queosk.config.baseEntity;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
