+++ queosk/src/main/java/com/bttf/queosk/config/swagger/SwaggerConfiguration.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.config;
+package com.bttf.queosk.config.swagger;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
