+++ queosk/src/main/java/com/bttf/queosk/entity/Order.java
@@ -1,6 +1,6 @@
 package com.bttf.queosk.entity;
 
-import com.bttf.queosk.config.BaseTimeEntity;
+import com.bttf.queosk.config.baseEntity.BaseTimeEntity;
 import com.bttf.queosk.dto.OrderStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
